+++ aoe2/aoe2_build_order.py
@@ -51,6 +51,10 @@ def check_valid_aoe2_build_order(data: dict) -> bool:
                 print(f'Build order \'{name}\' resources are not valid.')
                 return False
 
+            if ('builder' in resources) and (not is_valid_resource(resources['builder'])):  # optional builder count
+                print(f'Build order \'{name}\' builder resource is not valid.')
+                return False
+
             # notes
             notes = item['notes']
             for note in notes: