+++ Changelog.md
@@ -2,6 +2,8 @@
 * Hotkeys can be updated by reloading the settings.
 * Hotkeys can be set with a dedicated panel, without going in the settings file.
 * Build orders can be added in a dedicated panel, without having to create a file manually.
+* AoE2
+    * Builders can be indicated in the resources (optional).
 
 ## [1.3.4] - 2022.09.11
 * Resource targets can be split between different resource subtypes, by hovering the mouse.